backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
colorama==0.4.6
coverage==7.13.4
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==40.5.1
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==11.1.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-decouple==3.8
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
